rezaid
Innovation and staying ahead of the curve is the key to success when it comes to doing business. But we understand that for many growing companies it can be hard to source the technical talent needed to constantly innovate given the current skills shortage the sector is facing.

However Asian countries that are experiencing a technology boom have a glut of high quality software engineers, many of whom have experience of working abroad.

So, how can fast growing, innovative and ambitious SME’s take advantage of this talent when it’s based overseas?

That’s where we come in. Based in Manchester, Rezaid (UK) forms one half of an exclusive partnership with its sister company, Rezaid (Asia). Together, they provide bespoke software development and digital marketing services to close the productivity gap of high growth SMEs. Sharing a single ownership, both halves of the partnership benefit from a common operational structure. Our unique in-shoring model.

https://rezaid.co.uk/